api/requirements.txt
tensorflow==2.1.0
flask
pillow
scikit-learn==0.20.3
flask_cors
joblib
jupyter
matplotlib
Pillow
numpy
h5py
